Moved the service classes to respective sub-packages
