+++ src/main/java/in/gagan/base/framework/controller/AddressController.java
@@ -26,8 +26,7 @@ import in.gagan.base.framework.dto.location.ZipcodeDTO;
 import in.gagan.base.framework.entity.location.City;
 import in.gagan.base.framework.entity.location.Country;
 import in.gagan.base.framework.entity.location.Region;
-import in.gagan.base.framework.service.AddressService;
-import in.gagan.base.framework.util.DTOMapper;
+import in.gagan.base.framework.service.location.AddressService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
