+++ src/main/java/in/gagan/base/framework/service/user/PasswordManagerServiceImpl.java
@@ -17,14 +17,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package in.gagan.base.framework.service;
+package in.gagan.base.framework.service.user;
 
 import javax.transaction.Transactional;
 
+import in.gagan.base.framework.service.EmailService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import in.gagan.base.framework.dto.PasswordResetDTO;
 import in.gagan.base.framework.entity.ForgotPasswordToken;
 import in.gagan.base.framework.entity.User;
