+++ src/main/java/in/gagan/base/framework/service/user/UserRegisterationServiceImpl.java
@@ -17,10 +17,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package in.gagan.base.framework.service;
+package in.gagan.base.framework.service.user;
 
 import javax.transaction.Transactional;
 
+import in.gagan.base.framework.service.EmailService;
 import in.gagan.base.framework.util.DTOMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
